4.21 done

+++ controllers/blogs.js
@@ -18,8 +18,23 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blogToBedeleted = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if(blogToBedeleted.user.toString() === user._id.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(400).json({ error: 'blog is posted by another user' })
+  }
+
 })
 
 blogsRouter.post('/', async (request, response) => {
